수정: Revert 복구

+++ src/main/java/com/weet/app/user/domain/TrainerVO.java
@@ -11,6 +11,6 @@ public class TrainerVO {		// T_TR
 	private Integer userLevel;		// 레벨
 	private String userCareer;		// 경력
 	private String userIntro;		// 소개
-	private Integer userBiz;		// 사업자정보 
+	private Long userBiz;		// 사업자정보 
 	
 } // EmployeeVO
